add back navigation and adjust the height of the extension
