+++ web-x/package.json
@@ -18,6 +18,7 @@
     "jest-environment-jsdom": "^29.7.0",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
+    "react-icons": "^5.3.0",
     "react-redux": "^9.1.2",
     "react-router-dom": "^6.26.1",
     "sass": "^1.78.0"
