+++ web-x/src/pages/Home/Home.module.scss
@@ -13,7 +13,7 @@
 
     .body{
         // padding-top: rem(200);
-        min-height: rem(650);
+        height: rem(570);
         width: rem(350);
         background-color: var(--body-background);
         display: flex;
