+++ web-x/src/pages/Login/Login.module.scss
@@ -13,7 +13,7 @@
 
     .body{
         // padding-top: rem(200);
-        min-height: rem(650);
+        height: rem(570);
         width: rem(350);
         background-color: var(--body-background);
         display: flex;
@@ -21,7 +21,11 @@
         justify-content: center;
         align-items: center;
         gap: 0.5rem;
+        position: relative;
+    }
 
+    .back-icon{
+        @extend %back-click;
     }
 
     .onboarding-title{
