+++ web-x/src/pages/SeedPhrase/SeedPhraseConfirm.module.scss
@@ -11,7 +11,7 @@
     flex-wrap: wrap; 
     width: 80%; 
     gap: 10px;
-    margin-top: 2rem;
+    margin-top: 1.2rem;
     justify-content: center;
 }
 
@@ -22,7 +22,7 @@
     align-items: center;
     position: relative;
     .body{
-        height: rem(650);
+        height: rem(570);
         width: rem(350);
         background-color: var(--body-background);
         display: flex;
@@ -31,6 +31,11 @@
         align-items: center;
         position: relative;
     }
+
+    .back-icon{
+        @extend %back-click;
+    }
+
 
     .header{
 
@@ -74,7 +79,7 @@
 
         position: absolute;
         // top: 0.8rem;
-        bottom: 3rem;
+        bottom: 1.5rem;
         left: 1.2rem;
         margin-left: auto;
         margin-right: auto;
